engine: extract dispatch::encounter (#159)
