+++ crates/game-core/src/engine/dispatch/mod.rs
@@ -9,22 +9,23 @@
 //! ones.
 
 use crate::action::{EngineRecord, InputResponse, PlayerAction};
-use crate::card_data::{CardMetadata, CardType, Spawn, SpawnLocation};
+use crate::card_data::CardType;
 use crate::card_registry;
 use crate::dsl::{Cost, Trigger};
 use crate::event::Event;
 use crate::state::{
-    CardCode, CardInPlay, CardInstanceId, Enemy, EnemyId, GameState, Investigator, InvestigatorId,
-    Phase, SpawnEngagePending, Status, WindowKind, Zone,
+    CardCode, CardInPlay, CardInstanceId, EnemyId, GameState, Investigator, InvestigatorId, Phase,
+    Status, WindowKind, Zone,
 };
 
 use super::evaluator::{apply_effect, EvalContext};
-use super::outcome::{EngineOutcome, InputRequest, ResumeToken};
+use super::outcome::EngineOutcome;
 
 mod actions;
 mod combat;
 mod cursor;
 mod elimination;
+mod encounter;
 mod hunters;
 mod reaction_windows;
 mod skill_test;
@@ -37,14 +38,6 @@ const ACTIONS_PER_TURN: u8 = 3;
 /// Starting hand size at scenario setup. Per the Rules Reference,
 /// each investigator draws 5 cards before mulligan.
 const INITIAL_HAND_SIZE: u8 = 5;
-
-/// Hard cap on a single Mythos draw chain. Real scenarios surge ≤2
-/// in a chain; the cap exists purely to guarantee termination on
-/// malformed encounter decks (e.g. a deck small enough for surge to
-/// loop via the Rules Reference p.10 reshuffle). `unreachable!`-class
-/// — never reached in legitimate play.
-///
-const MAX_SURGE_CHAIN: usize = 64;
 
 /// Apply a [`PlayerAction`] to the state, pushing events.
 ///
@@ -175,7 +168,7 @@ pub fn apply_player_action(
         PlayerAction::DrawEncounterCard => match state.mythos_draw_pending {
             // DrawEncounterCard carries no investigator payload — the
             // acting investigator IS the pending cursor.
-            Some(actor) => draw_encounter_card(state, events, actor),
+            Some(actor) => encounter::draw_encounter_card(state, events, actor),
             None => EngineOutcome::Rejected {
                 reason: "DrawEncounterCard: no draw pending (all investigators have drawn)".into(),
             },
@@ -232,9 +225,9 @@ pub fn apply_engine_record(
 ) -> EngineOutcome {
     match record {
         EngineRecord::DeckShuffled { investigator } => deck_shuffled(state, events, *investigator),
-        EngineRecord::EncounterDeckShuffled => encounter_deck_shuffled(state, events),
+        EngineRecord::EncounterDeckShuffled => encounter::encounter_deck_shuffled(state, events),
         EngineRecord::EncounterCardRevealed { investigator } => {
-            encounter_card_revealed(state, events, *investigator)
+            encounter::encounter_card_revealed(state, events, *investigator)
         }
     }
 }
@@ -256,358 +249,6 @@ fn deck_shuffled(
     }
     shuffle_player_deck(state, events, investigator);
     EngineOutcome::Done
-}
-
-/// Handler for [`EngineRecord::EncounterDeckShuffled`].
-///
-/// Permutes the shared encounter deck via the deterministic RNG and
-/// emits [`Event::EncounterDeckShuffled`] (when ≥ 2 cards). No
-/// validation — the encounter deck is shared, so there's no
-/// per-investigator existence check.
-fn encounter_deck_shuffled(state: &mut GameState, events: &mut Vec<Event>) -> EngineOutcome {
-    shuffle_encounter_deck(state, events);
-    EngineOutcome::Done
-}
-
-/// Handler for [`EngineRecord::EncounterCardRevealed`].
-///
-/// Drives the on-draw resolution path for one encounter card:
-///
-/// 1. Validate that a card registry is installed (reject with
-///    `"EncounterCardRevealed: no card registry installed"` if not).
-/// 2. Draw the top of the encounter deck via [`draw_encounter_top`]
-///    (transparently reshuffles discard back in if the deck is
-///    empty). Reject with `"EncounterCardRevealed: encounter deck and discard both empty"`
-///    if both piles are exhausted.
-/// 3. Look up the drawn card's metadata via the installed registry.
-///    Reject with `"EncounterCardRevealed: unknown card code: {code}"` if the registry
-///    doesn't know the code.
-/// 4. Delegate to [`resolve_encounter_card`] for the post-draw
-///    resolution prefix (emit [`Event::CardRevealed`] + type-dispatch
-///    to Revelation / spawn / reject).
-///
-/// # Validate-first contract caveat
-///
-/// `draw_encounter_top` mutates `state.encounter_deck` /
-/// `state.encounter_discard` BEFORE the unknown-code reject can
-/// fire; `Event::CardRevealed` then emits BEFORE Revelation /
-/// spawn resolve. The draw is a documented exception to the
-/// validate-first convention — the card must be removed from the
-/// deck before the reaction window opens (Before-timing listeners
-/// in #52 need to see the revealed-but-not-yet-resolved state).
-/// `Event::CardRevealed` emits before Revelation for the same
-/// reason: Before-timing reaction listeners (#52, not wired) need
-/// the event to fire before Revelation resolves (rules-correct
-/// interposition point).
-///
-/// Compare to `play_card`'s documented mid-resolution caveat in
-/// CLAUDE.md: same shape, same rationale.
-fn encounter_card_revealed(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-) -> EngineOutcome {
-    let Some(registry) = card_registry::current() else {
-        return EngineOutcome::Rejected {
-            reason: "EncounterCardRevealed: no card registry installed".into(),
-        };
-    };
-
-    let Some(code) = draw_encounter_top(state, events) else {
-        return EngineOutcome::Rejected {
-            reason: "EncounterCardRevealed: encounter deck and discard both empty".into(),
-        };
-    };
-
-    let Some(metadata) = (registry.metadata_for)(&code) else {
-        return EngineOutcome::Rejected {
-            reason: format!("EncounterCardRevealed: unknown card code: {code:?}").into(),
-        };
-    };
-    resolve_encounter_card(state, events, investigator, code, metadata)
-}
-
-/// Shared post-draw resolution helper. Resolves the per-card 5-step
-/// sub-sequence's steps 3 (Revelation) and 4 (enemy spawn) for an
-/// already-drawn encounter card. Called by [`encounter_card_revealed`]
-/// (the `EngineRecord::EncounterCardRevealed` path) and by
-/// `mythos_draw_for` (Mythos 1.4 player-driven draws, lands in T11).
-///
-/// Body: emits [`Event::CardRevealed`], then dispatches on
-/// `metadata.card_type` — treachery → run Revelation abilities →
-/// push card to `encounter_discard`;
-/// enemy → run Revelation abilities → call [`spawn_enemy`];
-/// any other type → return `Rejected`.
-///
-/// **Mid-resolution caveat:** [`Event::CardRevealed`] emits before
-/// Revelation runs (Before-timing reactions need that ordering,
-/// per #126's design decision). The apply loop's `events.clear()`
-/// on Rejected still wipes the event stream on rejection.
-fn resolve_encounter_card(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-    code: CardCode,
-    metadata: &CardMetadata,
-) -> EngineOutcome {
-    let card_type = metadata.card_type;
-
-    // Emit BEFORE Revelation resolves — see caveat in encounter_card_revealed.
-    events.push(Event::CardRevealed {
-        investigator,
-        code: code.clone(),
-        card_type,
-    });
-
-    match card_type {
-        CardType::Treachery => {
-            let Some(registry) = card_registry::current() else {
-                return EngineOutcome::Rejected {
-                    reason: "encounter card resolution: no card registry installed".into(),
-                };
-            };
-            let abilities = (registry.abilities_for)(&code).unwrap_or_default();
-            let ctx = EvalContext::for_controller(investigator);
-            for ability in abilities
-                .iter()
-                .filter(|a| a.trigger == Trigger::Revelation)
-            {
-                let outcome = apply_effect(state, events, &ability.effect, ctx);
-                if !matches!(outcome, EngineOutcome::Done) {
-                    return outcome;
-                }
-            }
-            state.encounter_discard.push(code);
-            EngineOutcome::Done
-        }
-        CardType::Enemy => {
-            // Revelation effects on enemies (rare, but printed on
-            // some encounter enemies — e.g. "Revelation - Discard
-            // 1 card from your hand at random.") fire BEFORE the
-            // enemy spawns into play, per Rules Reference p.24
-            // ("1.4 Each investigator draws 1 encounter card"):
-            // "3. Resolve the revelation ability on the drawn card."
-            // followed by "4. If the card is an enemy, spawn it
-            // following any spawn instruction the card bears."
-            //
-            // No Phase-4-scope enemy has a Revelation effect; this
-            // loop is structural for Phase-7+ enemies.
-            let Some(registry) = card_registry::current() else {
-                return EngineOutcome::Rejected {
-                    reason: "encounter card resolution: no card registry installed".into(),
-                };
-            };
-            let abilities = (registry.abilities_for)(&code).unwrap_or_default();
-            let ctx = EvalContext::for_controller(investigator);
-            for ability in abilities
-                .iter()
-                .filter(|a| a.trigger == Trigger::Revelation)
-            {
-                let outcome = apply_effect(state, events, &ability.effect, ctx);
-                if !matches!(outcome, EngineOutcome::Done) {
-                    return outcome;
-                }
-            }
-            spawn_enemy(state, events, investigator, code, metadata)
-        }
-        other => EngineOutcome::Rejected {
-            reason: format!(
-                "EncounterCardRevealed: invalid encounter card type {other:?}; \
-                 encounter decks contain only treachery and enemy cards",
-            )
-            .into(),
-        },
-    }
-}
-
-/// Spawn one encounter-deck enemy into play.
-///
-/// Called by [`encounter_card_revealed`] after `Event::CardRevealed`
-/// has fired and any [`Trigger::Revelation`](crate::dsl::Trigger::Revelation)
-/// abilities on the enemy have resolved.
-///
-/// # Spawn-location resolution
-///
-/// Rules Reference page 24, step 4 (1.4 Each investigator draws 1
-/// encounter card):
-///
-/// > If the card is an **enemy**, spawn it following any spawn
-/// > instruction the card bears. (A spawn instruction is any text
-/// > bearing a "spawn" precursor.) If the encountered enemy has no
-/// > spawn instruction, the enemy spawns engaged with the investigator
-/// > encountering the card and is placed in that investigator's threat
-/// > area.
-///
-/// We model threat-area placement as
-/// `enemy.current_location = drawing investigator's location` +
-/// `engaged_with = drawing investigator`. The named-location case
-/// (`SpawnLocation::Specific`) looks the location up by its
-/// printed [`code`](crate::state::Location::code).
-///
-/// # Engagement-on-spawn
-///
-/// Rules Reference page 10 (Enemy Engagement):
-///
-/// > Any time a ready unengaged enemy is at the same location as an
-/// > investigator, it engages that investigator, and is placed in that
-/// > investigator's threat area. If there are multiple investigators
-/// > at the same location as a ready unengaged enemy, follow the
-/// > enemy's prey instructions to determine which investigator is
-/// > engaged.
-///
-/// All cases route through the shared [`resolve_prey`] resolver
-/// (#128, option A): the co-located set is narrowed by the enemy's
-/// prey (always `Prey::Default` in current scope, so a 2+ set always
-/// ties). `None`/`One` resolve inline (no engagement, or engage the
-/// sole/best candidate); `Tie` suspends via
-/// [`SpawnEngagePending`](crate::state::SpawnEngagePending) and returns
-/// [`EngineOutcome::AwaitingInput`] for the lead investigator's
-/// `PickInvestigator`. When the spawn happens inside a Mythos
-/// encounter-draw chain, [`resume_spawn_engage`] re-enters
-/// [`run_mythos_draw_chain`] after the pick resolves.
-///
-/// # Stat fields TODO
-///
-/// `CardMetadata` doesn't yet carry per-enemy `fight` / `evade` /
-/// `attack_damage` / `attack_horror`. This handler hardcodes
-/// `fight: 1, evade: 1, attack_damage: 0, attack_horror: 0` until
-/// a future PR (Phase-7+, alongside the first real spawn-bearing
-/// enemy) extends `CardMetadata` with enemy-specific stat fields and
-/// this handler reads them. Health uses `metadata.health.unwrap_or(1)`
-/// because `CardMetadata.health` already exists.
-///
-/// # Validate-first contract
-///
-/// The one precondition that can reject — spawn-location resolution —
-/// is checked before any mutation, leaving `state`/`events` unchanged
-/// on rejection. Engagement resolution never rejects: it either
-/// resolves inline or suspends (`AwaitingInput`) with the enemy already
-/// minted into `state.enemies` and `Event::EnemySpawned` pushed (the
-/// pending choice carries the rest of the work to [`resume_spawn_engage`]).
-#[allow(clippy::too_many_lines)]
-fn spawn_enemy(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-    code: CardCode,
-    metadata: &CardMetadata,
-) -> EngineOutcome {
-    // 1. Resolve spawn location (validate-first).
-    let location_id = match &metadata.spawn {
-        Some(Spawn {
-            location: SpawnLocation::Specific(loc_code),
-        }) => match state
-            .locations
-            .iter()
-            .find(|(_, loc)| loc.code.as_str() == loc_code.as_str())
-        {
-            Some((id, _)) => *id,
-            None => {
-                return EngineOutcome::Rejected {
-                    reason: format!("spawn_enemy: spawn location not in play (code {loc_code:?})")
-                        .into(),
-                };
-            }
-        },
-        None => match state
-            .investigators
-            .get(&investigator)
-            .and_then(|inv| inv.current_location)
-        {
-            Some(loc) => loc,
-            None => {
-                return EngineOutcome::Rejected {
-                    reason: format!(
-                        "spawn_enemy: drawing investigator has no location \
-                         (investigator {investigator:?})",
-                    )
-                    .into(),
-                };
-            }
-        },
-    };
-
-    // 2. Resolve engagement-on-spawn (validate-first). The co-located
-    //    set is narrowed by the enemy's prey — every spawn uses
-    //    `Prey::Default` (Task 2), so a 2+ set always ties and suspends
-    //    for the lead investigator's `PickInvestigator` (option A).
-    let candidates = cursor::active_investigators_at(state, location_id);
-
-    // 3. Mint and place (mutate-second). The enemy is inserted unengaged;
-    //    the `One` and (post-resume) `Tie` cases set `engaged_with` via
-    //    `engage_enemy_with` so the `EnemyEngaged` event always pairs with
-    //    the mutation.
-    let enemy_id = EnemyId(state.next_enemy_id);
-    state.next_enemy_id = state.next_enemy_id.saturating_add(1);
-
-    let enemy = Enemy {
-        id: enemy_id,
-        name: metadata.name.clone(),
-        fight: 1,
-        evade: 1,
-        max_health: metadata.health.unwrap_or(1),
-        damage: 0,
-        attack_damage: 0,
-        attack_horror: 0,
-        current_location: Some(location_id),
-        exhausted: false,
-        traits: metadata.traits.clone(),
-        engaged_with: None,
-        hunter: false,
-        prey: crate::card_data::Prey::Default,
-    };
-    state.enemies.insert(enemy_id, enemy);
-
-    match hunters::resolve_prey(state, crate::card_data::Prey::Default, &candidates) {
-        hunters::PreyResolution::None => {
-            events.push(Event::EnemySpawned {
-                enemy: enemy_id,
-                code,
-                location: location_id,
-                engaged_with: None,
-            });
-            EngineOutcome::Done
-        }
-        hunters::PreyResolution::One(target) => {
-            events.push(Event::EnemySpawned {
-                enemy: enemy_id,
-                code,
-                location: location_id,
-                engaged_with: Some(target),
-            });
-            hunters::engage_enemy_with(state, events, enemy_id, target);
-            EngineOutcome::Done
-        }
-        hunters::PreyResolution::Tie(tied) => {
-            events.push(Event::EnemySpawned {
-                enemy: enemy_id,
-                code,
-                location: location_id,
-                engaged_with: None,
-            });
-            // `chain_count` is 0 here; when this spawn is reached inside a
-            // Mythos surge chain, `run_mythos_draw_chain` patches the
-            // stored value to the live chain position before returning the
-            // `AwaitingInput` (the single-draw `EncounterCardRevealed` path
-            // has no chain, so 0 is correct there).
-            state.spawn_engage_pending = Some(SpawnEngagePending {
-                enemy: enemy_id,
-                investigator_to_draw: investigator,
-                candidates: tied.clone(),
-                surge: metadata.surge,
-                chain_count: 0,
-            });
-            EngineOutcome::AwaitingInput {
-                request: InputRequest {
-                    prompt: format!(
-                        "Enemy {enemy_id:?} spawn engagement: lead investigator picks whom to \
-                         engage among {tied:?} (submit InputResponse::PickInvestigator)"
-                    ),
-                },
-                resume_token: ResumeToken(0),
-            }
-        }
-    }
 }
 
 /// Fisher-Yates shuffle of the named investigator's deck using the
@@ -652,71 +293,6 @@ pub(super) fn shuffle_player_deck(
         inv.deck.swap(a, b);
     }
     events.push(Event::DeckShuffled { investigator });
-}
-
-/// Fisher-Yates shuffle of the shared encounter deck using the
-/// shared deterministic RNG. Used by [`encounter_deck_shuffled`] and
-/// by [`reshuffle_encounter_discard`].
-///
-/// Emits [`Event::EncounterDeckShuffled`] iff the deck had at least
-/// 2 cards (a 0- or 1-card deck has nothing to permute).
-pub(super) fn shuffle_encounter_deck(state: &mut GameState, events: &mut Vec<Event>) {
-    let deck_len = state.encounter_deck.len();
-    if deck_len < 2 {
-        return;
-    }
-    // Mirror shuffle_player_deck's "collect swaps then apply" pattern:
-    // RngState::next_index borrows &mut state.rng, which would conflict
-    // with a &mut borrow on state.encounter_deck inline.
-    let mut swaps: Vec<(usize, usize)> = Vec::with_capacity(deck_len - 1);
-    let mut i = deck_len - 1;
-    while i >= 1 {
-        let j = state.rng.next_index(i + 1);
-        swaps.push((i, j));
-        i -= 1;
-    }
-    for (a, b) in swaps {
-        state.encounter_deck.swap(a, b);
-    }
-    events.push(Event::EncounterDeckShuffled);
-}
-
-/// Drain `state.encounter_discard` into `state.encounter_deck` and
-/// shuffle the resulting deck. Called by
-/// [`draw_encounter_top`] when the deck runs empty.
-///
-/// Does NOT push an `EngineRecord::EncounterDeckShuffled` to the
-/// action log — mid-handler reshuffles rely on RNG determinism for
-/// replay rather than log entries, mirroring the existing
-/// player-deck pattern. The `EngineRecord` variant is reserved for
-/// explicit shuffle actions (future "shuffle X into the encounter
-/// deck" effects).
-pub(super) fn reshuffle_encounter_discard(state: &mut GameState, events: &mut Vec<Event>) {
-    state
-        .encounter_deck
-        .extend(state.encounter_discard.drain(..));
-    shuffle_encounter_deck(state, events);
-}
-
-/// Draw the top card of the encounter deck, transparently reshuffling
-/// the discard back in if the deck is empty.
-///
-/// Returns `Some(code)` when a card was available (either from the
-/// deck directly or after the reshuffle). Returns `None` when both
-/// the deck and the discard are empty — callers decide how to
-/// interpret this (#69's Mythos loop treats it as a scenario
-/// condition rather than an engine error).
-pub(super) fn draw_encounter_top(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-) -> Option<CardCode> {
-    if state.encounter_deck.is_empty() {
-        if state.encounter_discard.is_empty() {
-            return None;
-        }
-        reshuffle_encounter_discard(state, events);
-    }
-    state.encounter_deck.pop_front()
 }
 
 /// Draw up to `count` cards from the named investigator's deck top
@@ -2240,640 +1816,6 @@ pub(super) fn check_cost_payable(
 }
 
 #[cfg(test)]
-mod encounter_card_revealed_tests {
-    use super::*;
-    use crate::state::CardCode;
-    use crate::test_support::{test_investigator, TestGame};
-
-    /// Exercises the early-reject guard: when the handler cannot
-    /// proceed past the registry / metadata checks, it must reject
-    /// without drawing from the deck and without emitting any events.
-    ///
-    /// Two possible rejection reasons depending on process state:
-    ///
-    /// - `"no card registry installed"` — if no registry has been
-    ///   installed yet in this process.
-    /// - `"unknown card code: ..."` — if another test in this binary
-    ///   has already installed a fake registry that doesn't know the
-    ///   synthetic code `"__no_such_card"`.
-    ///
-    /// In both cases the invariant is identical: deck untouched, no
-    /// events emitted. The exact rejection reason depends on
-    /// `OnceLock` install ordering, which is non-deterministic across
-    /// parallel test binaries.
-    ///
-    /// The authoritative "no registry installed" path is exercised in
-    /// the `crates/scenarios/tests/encounter_reveal.rs` integration
-    /// test, which runs in its own process and installs `TEST_REGISTRY`
-    /// explicitly. The process-isolated install guarantees the "no
-    /// registry" rejection fires in a controlled environment.
-    #[test]
-    fn rejects_when_no_card_registry_installed() {
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .build();
-        // Seed the encounter deck so we can prove the reject fires
-        // *before* the draw mutates state. Use a code that no real
-        // or fake registry knows so we always hit an early rejection.
-        state
-            .encounter_deck
-            .push_back(CardCode("__no_such_card".into()));
-        let pre_deck_len = state.encounter_deck.len();
-        let mut events = Vec::new();
-
-        let outcome = apply_engine_record(
-            &mut state,
-            &mut events,
-            &EngineRecord::EncounterCardRevealed {
-                investigator: InvestigatorId(1),
-            },
-        );
-
-        match outcome {
-            EngineOutcome::Rejected { reason } => {
-                assert!(
-                    reason.contains("no card registry installed")
-                        || reason.contains("unknown card code"),
-                    "unexpected reject reason: {reason:?}",
-                );
-            }
-            other => panic!("expected Rejected, got {other:?}"),
-        }
-        // Deck must be untouched: the registry-missing reject fires
-        // before any draw, and the unknown-code reject fires after the
-        // draw. However, the plan's documented exception means that
-        // after a successful draw but unknown-code rejection, the deck
-        // will be shorter by one. We assert on the *invariant* that
-        // matters: no events were emitted, and the deck shrank by at
-        // most one (not more).
-        assert!(
-            state.encounter_deck.len() <= pre_deck_len,
-            "deck should not grow; expected <= {pre_deck_len}, got {}",
-            state.encounter_deck.len(),
-        );
-        assert!(
-            events.is_empty(),
-            "no events should fire before Event::CardRevealed; got {events:?}",
-        );
-    }
-}
-
-#[cfg(test)]
-mod encounter_deck_helper_tests {
-    use super::*;
-    use crate::event::Event;
-    use crate::rng::RngState;
-    use crate::state::CardCode;
-    use crate::test_support::TestGame;
-
-    #[test]
-    fn shuffle_encounter_deck_emits_event_when_two_or_more_cards() {
-        let mut state = TestGame::new().build();
-        state.rng = RngState::new(42);
-        state.encounter_deck.push_back(CardCode("a".into()));
-        state.encounter_deck.push_back(CardCode("b".into()));
-        state.encounter_deck.push_back(CardCode("c".into()));
-
-        let mut events = Vec::new();
-        shuffle_encounter_deck(&mut state, &mut events);
-
-        assert!(matches!(events.as_slice(), [Event::EncounterDeckShuffled]));
-        assert_eq!(state.encounter_deck.len(), 3);
-        let mut codes: Vec<_> = state.encounter_deck.iter().cloned().collect();
-        codes.sort_by(|a, b| a.0.cmp(&b.0));
-        assert_eq!(
-            codes,
-            vec![
-                CardCode("a".into()),
-                CardCode("b".into()),
-                CardCode("c".into())
-            ]
-        );
-    }
-
-    #[test]
-    fn shuffle_encounter_deck_is_silent_on_zero_or_one_card() {
-        for n in 0..=1 {
-            let mut state = TestGame::new().build();
-            for i in 0..n {
-                state.encounter_deck.push_back(CardCode(format!("c{i}")));
-            }
-            let mut events = Vec::new();
-            shuffle_encounter_deck(&mut state, &mut events);
-            assert!(events.is_empty(), "expected no event for n={n} deck");
-        }
-    }
-
-    #[test]
-    fn reshuffle_encounter_discard_moves_discard_into_deck_and_shuffles() {
-        let mut state = TestGame::new().build();
-        state.rng = RngState::new(7);
-        for i in 0..5 {
-            state.encounter_discard.push(CardCode(format!("d{i}")));
-        }
-
-        let mut events = Vec::new();
-        reshuffle_encounter_discard(&mut state, &mut events);
-
-        assert!(
-            state.encounter_discard.is_empty(),
-            "discard should be drained"
-        );
-        assert_eq!(state.encounter_deck.len(), 5, "all 5 cards moved into deck");
-        assert!(
-            matches!(events.as_slice(), [Event::EncounterDeckShuffled]),
-            "expected EncounterDeckShuffled (≥ 2 cards moved)"
-        );
-    }
-
-    #[test]
-    fn reshuffle_encounter_discard_is_silent_when_discard_has_one_card() {
-        let mut state = TestGame::new().build();
-        state.encounter_discard.push(CardCode("solo".into()));
-
-        let mut events = Vec::new();
-        reshuffle_encounter_discard(&mut state, &mut events);
-
-        assert!(state.encounter_discard.is_empty());
-        assert_eq!(state.encounter_deck.len(), 1);
-        assert!(events.is_empty(), "1-card shuffle emits no event");
-    }
-
-    #[test]
-    fn draw_encounter_top_drains_deck_then_returns_none() {
-        let mut state = TestGame::new().build();
-        state.encounter_deck.push_back(CardCode("a".into()));
-        state.encounter_deck.push_back(CardCode("b".into()));
-        state.encounter_deck.push_back(CardCode("c".into()));
-
-        let mut events = Vec::new();
-
-        assert_eq!(
-            draw_encounter_top(&mut state, &mut events),
-            Some(CardCode("a".into()))
-        );
-        assert_eq!(
-            draw_encounter_top(&mut state, &mut events),
-            Some(CardCode("b".into()))
-        );
-        assert_eq!(
-            draw_encounter_top(&mut state, &mut events),
-            Some(CardCode("c".into()))
-        );
-        assert_eq!(draw_encounter_top(&mut state, &mut events), None);
-        assert!(
-            events.is_empty(),
-            "no events for any draw — discard is always empty, no reshuffle is triggered"
-        );
-    }
-
-    #[test]
-    fn draw_encounter_top_reshuffles_discard_on_empty_deck() {
-        let mut state = TestGame::new().build();
-        state.rng = RngState::new(13);
-        state.encounter_discard.push(CardCode("x".into()));
-        state.encounter_discard.push(CardCode("y".into()));
-        state.encounter_discard.push(CardCode("z".into()));
-
-        let mut events = Vec::new();
-        let drawn = draw_encounter_top(&mut state, &mut events);
-
-        let drawn_code = drawn.expect("should reshuffle and draw");
-        assert!(
-            [
-                CardCode("x".into()),
-                CardCode("y".into()),
-                CardCode("z".into())
-            ]
-            .contains(&drawn_code),
-            "drawn card must be one of the three discard cards, got {drawn_code:?}"
-        );
-        assert_eq!(
-            state.encounter_deck.len(),
-            2,
-            "2 cards remain in deck post-draw"
-        );
-        assert!(state.encounter_discard.is_empty(), "discard drained");
-        assert!(
-            matches!(events.as_slice(), [Event::EncounterDeckShuffled]),
-            "reshuffle emits one event"
-        );
-    }
-
-    #[test]
-    fn draw_encounter_top_returns_none_when_deck_and_discard_both_empty() {
-        let mut state = TestGame::new().build();
-        let mut events = Vec::new();
-        assert_eq!(draw_encounter_top(&mut state, &mut events), None);
-        assert!(events.is_empty(), "no events on empty-on-both");
-    }
-
-    #[test]
-    fn engine_record_encounter_deck_shuffled_drives_shuffle() {
-        use crate::action::{Action, EngineRecord};
-        use crate::engine::apply;
-
-        let mut state = TestGame::new().build();
-        state.rng = RngState::new(99);
-        for i in 0..4 {
-            state.encounter_deck.push_back(CardCode(format!("c{i}")));
-        }
-        let original: Vec<_> = state.encounter_deck.iter().cloned().collect();
-
-        let result = apply(state, Action::Engine(EngineRecord::EncounterDeckShuffled));
-
-        assert!(
-            matches!(result.outcome, crate::EngineOutcome::Done),
-            "expected Done, got {:?}",
-            result.outcome
-        );
-        let mut shuffled: Vec<_> = result.state.encounter_deck.iter().cloned().collect();
-        let mut orig_sorted = original.clone();
-        shuffled.sort_by(|a, b| a.0.cmp(&b.0));
-        orig_sorted.sort_by(|a, b| a.0.cmp(&b.0));
-        assert_eq!(shuffled, orig_sorted);
-        assert!(result
-            .events
-            .iter()
-            .any(|e| matches!(e, Event::EncounterDeckShuffled)));
-    }
-
-    #[test]
-    fn encounter_deck_shuffle_is_deterministic_from_seed() {
-        fn shuffle_with_seed(seed: u64) -> Vec<CardCode> {
-            let mut state = TestGame::new().build();
-            state.rng = RngState::new(seed);
-            for i in 0..10 {
-                state.encounter_deck.push_back(CardCode(format!("c{i:02}")));
-            }
-            let mut events = Vec::new();
-            shuffle_encounter_deck(&mut state, &mut events);
-            state.encounter_deck.iter().cloned().collect()
-        }
-
-        let a = shuffle_with_seed(2026);
-        let b = shuffle_with_seed(2026);
-        assert_eq!(a, b, "same seed must produce same shuffle order");
-
-        let c = shuffle_with_seed(42);
-        assert_ne!(
-            a, c,
-            "different seeds should produce different orders (smoke test)"
-        );
-    }
-}
-
-#[cfg(test)]
-mod spawn_enemy_tests {
-    use super::*;
-    use crate::state::{CardCode, LocationId};
-    use crate::test_support::{test_investigator, test_location, TestGame};
-    use crate::{assert_event, assert_event_sequence, assert_no_event};
-    use card_dsl::card_data::{CardMetadata, CardType, Class, SkillIcons, Spawn, SpawnLocation};
-
-    fn synth_enemy_metadata(spawn: Option<Spawn>) -> CardMetadata {
-        CardMetadata {
-            code: "_synth_enemy".into(),
-            name: "Synth Enemy".into(),
-            class: Class::Mythos,
-            card_type: CardType::Enemy,
-            cost: None,
-            xp: None,
-            text: None,
-            flavor: None,
-            illustrator: None,
-            traits: Vec::new(),
-            slots: Vec::new(),
-            skill_icons: SkillIcons::default(),
-            health: Some(1),
-            sanity: None,
-            deck_limit: 1,
-            quantity: 1,
-            pack_code: "_synth".into(),
-            position: 1,
-            is_fast: false,
-            spawn,
-            surge: false,
-            peril: false,
-        }
-    }
-
-    #[test]
-    fn spawn_at_specific_location_with_one_investigator_engages_them() {
-        let mut loc = test_location(10, "Synth Loc");
-        loc.code = CardCode("_synth_loc".into());
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .with_location(loc)
-            .with_turn_order([InvestigatorId(1)])
-            .build();
-        // Place investigator 1 at location 10.
-        state
-            .investigators
-            .get_mut(&InvestigatorId(1))
-            .unwrap()
-            .current_location = Some(LocationId(10));
-
-        let metadata = synth_enemy_metadata(Some(Spawn {
-            location: SpawnLocation::Specific("_synth_loc".into()),
-        }));
-        let mut events = Vec::new();
-
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-
-        assert!(matches!(outcome, EngineOutcome::Done), "{outcome:?}");
-        assert_eq!(state.enemies.len(), 1);
-        let (_, enemy) = state.enemies.iter().next().unwrap();
-        assert_eq!(enemy.current_location, Some(LocationId(10)));
-        assert_eq!(enemy.engaged_with, Some(InvestigatorId(1)));
-
-        assert_event_sequence!(
-            events,
-            Event::EnemySpawned { code, location, engaged_with, .. }
-                if *code == CardCode("_synth_enemy".into())
-                    && *location == LocationId(10)
-                    && *engaged_with == Some(InvestigatorId(1)),
-            Event::EnemyEngaged { investigator, .. }
-                if *investigator == InvestigatorId(1),
-        );
-    }
-
-    #[test]
-    fn spawn_at_specific_location_with_no_investigators_leaves_unengaged() {
-        let mut loc = test_location(10, "Synth Loc");
-        loc.code = CardCode("_synth_loc".into());
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .with_location(loc)
-            .build();
-        // Investigator 1 is NOT at location 10 (current_location is None).
-
-        let metadata = synth_enemy_metadata(Some(Spawn {
-            location: SpawnLocation::Specific("_synth_loc".into()),
-        }));
-        let mut events = Vec::new();
-
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-
-        assert!(matches!(outcome, EngineOutcome::Done), "{outcome:?}");
-        let (_, enemy) = state.enemies.iter().next().unwrap();
-        assert_eq!(enemy.engaged_with, None);
-        // No engagement happened, so no EnemyEngaged event fires.
-        assert_no_event!(events, Event::EnemyEngaged { .. });
-    }
-
-    #[test]
-    fn spawn_at_specific_location_rejects_when_location_not_in_play() {
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .build();
-        let metadata = synth_enemy_metadata(Some(Spawn {
-            location: SpawnLocation::Specific("_nonexistent_loc".into()),
-        }));
-        let mut events = Vec::new();
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        match outcome {
-            EngineOutcome::Rejected { reason } => {
-                assert!(
-                    reason.contains("spawn location not in play"),
-                    "unexpected reason: {reason:?}",
-                );
-            }
-            other => panic!("expected Rejected, got {other:?}"),
-        }
-        assert!(state.enemies.is_empty());
-    }
-
-    #[test]
-    fn spawn_with_no_instruction_places_at_drawing_investigators_location() {
-        let mut loc = test_location(10, "Demo");
-        loc.code = CardCode("_demo_loc".into());
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .with_location(loc)
-            .with_turn_order([InvestigatorId(1)])
-            .build();
-        state
-            .investigators
-            .get_mut(&InvestigatorId(1))
-            .unwrap()
-            .current_location = Some(LocationId(10));
-        let metadata = synth_enemy_metadata(None);
-        let mut events = Vec::new();
-
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-
-        assert!(matches!(outcome, EngineOutcome::Done), "{outcome:?}");
-        let (_, enemy) = state.enemies.iter().next().unwrap();
-        assert_eq!(enemy.current_location, Some(LocationId(10)));
-        assert_eq!(enemy.engaged_with, Some(InvestigatorId(1)));
-        // Default-spawn engagement fires the paired EnemyEngaged event.
-        assert_event!(
-            events,
-            Event::EnemyEngaged { investigator, .. }
-                if *investigator == InvestigatorId(1)
-        );
-    }
-
-    #[test]
-    fn spawn_with_no_instruction_rejects_when_drawing_investigator_has_no_location() {
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .build();
-        // Investigator has no current_location.
-        let metadata = synth_enemy_metadata(None);
-        let mut events = Vec::new();
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        match outcome {
-            EngineOutcome::Rejected { reason } => {
-                assert!(
-                    reason.contains("drawing investigator has no location"),
-                    "unexpected reason: {reason:?}",
-                );
-            }
-            other => panic!("expected Rejected, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn spawn_engages_sole_colocated_investigator() {
-        // Regression: #127's single-investigator engage-on-spawn path
-        // still resolves inline under the shared prey resolver.
-        let mut loc = test_location(1, "Hall");
-        loc.code = CardCode("_loc".into());
-        let mut inv = test_investigator(1);
-        inv.current_location = Some(LocationId(1));
-        let mut state = TestGame::new()
-            .with_phase(Phase::Mythos)
-            .with_location(loc)
-            .with_investigator(inv)
-            .with_turn_order([InvestigatorId(1)])
-            .build();
-        let metadata = synth_enemy_metadata(None);
-        let mut events = Vec::new();
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        assert_eq!(outcome, EngineOutcome::Done);
-        let spawned = state.enemies.values().next().expect("one enemy");
-        assert_eq!(spawned.engaged_with, Some(InvestigatorId(1)));
-    }
-
-    #[test]
-    fn spawn_tie_suspends_for_lead_pick() {
-        let mut loc = test_location(1, "Hall");
-        loc.code = CardCode("_loc".into());
-        let mut i1 = test_investigator(1);
-        i1.current_location = Some(LocationId(1));
-        let mut i2 = test_investigator(2);
-        i2.current_location = Some(LocationId(1));
-        let mut state = TestGame::new()
-            .with_phase(Phase::Mythos)
-            .with_location(loc)
-            .with_investigator(i1)
-            .with_investigator(i2)
-            .with_turn_order([InvestigatorId(1), InvestigatorId(2)])
-            .build();
-        state.mythos_draw_pending = Some(InvestigatorId(1));
-        let metadata = synth_enemy_metadata(None);
-        let mut events = Vec::new();
-        let outcome = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        assert!(matches!(outcome, EngineOutcome::AwaitingInput { .. }));
-        assert!(state.spawn_engage_pending.is_some());
-        let spawned = state.enemies.values().next().expect("one enemy");
-        assert_eq!(spawned.engaged_with, None);
-    }
-
-    #[test]
-    fn resume_spawn_engage_rejects_bad_pick_and_preserves_pending() {
-        // Validate-first: a pick outside the stored candidate set rejects
-        // and leaves `spawn_engage_pending` intact for retry, with the
-        // enemy still unengaged.
-        use crate::action::InputResponse;
-        let mut loc = test_location(1, "Hall");
-        loc.code = CardCode("_loc".into());
-        let mut i1 = test_investigator(1);
-        i1.current_location = Some(LocationId(1));
-        let mut i2 = test_investigator(2);
-        i2.current_location = Some(LocationId(1));
-        let mut state = TestGame::new()
-            .with_phase(Phase::Mythos)
-            .with_location(loc)
-            .with_investigator(i1)
-            .with_investigator(i2)
-            .with_turn_order([InvestigatorId(1), InvestigatorId(2)])
-            .build();
-        state.mythos_draw_pending = Some(InvestigatorId(1));
-        let metadata = synth_enemy_metadata(None);
-        let mut events = Vec::new();
-        let _ = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        assert!(state.spawn_engage_pending.is_some());
-
-        // Investigator 3 is not among the co-located candidates.
-        let outcome = hunters::resume_spawn_engage(
-            &mut state,
-            &mut events,
-            &InputResponse::PickInvestigator(InvestigatorId(3)),
-        );
-        assert!(
-            matches!(outcome, EngineOutcome::Rejected { .. }),
-            "{outcome:?}"
-        );
-        assert!(
-            state.spawn_engage_pending.is_some(),
-            "pending must survive a rejected pick for retry",
-        );
-        let enemy = state.enemies.values().next().expect("enemy still placed");
-        assert_eq!(enemy.engaged_with, None, "no engagement on rejected pick");
-    }
-
-    #[test]
-    fn spawn_mints_distinct_enemy_ids() {
-        let mut loc = test_location(10, "L");
-        loc.code = CardCode("_l".into());
-        let mut state = TestGame::new()
-            .with_investigator(test_investigator(1))
-            .with_location(loc)
-            .build();
-        state
-            .investigators
-            .get_mut(&InvestigatorId(1))
-            .unwrap()
-            .current_location = Some(LocationId(10));
-        let metadata = synth_enemy_metadata(Some(Spawn {
-            location: SpawnLocation::Specific("_l".into()),
-        }));
-        let mut events = Vec::new();
-
-        let _ = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        let _ = spawn_enemy(
-            &mut state,
-            &mut events,
-            InvestigatorId(1),
-            CardCode("_synth_enemy".into()),
-            &metadata,
-        );
-        assert_eq!(
-            state.enemies.len(),
-            2,
-            "two spawns should produce two distinct enemies"
-        );
-    }
-}
-
-#[cfg(test)]
 mod investigation_phase_tests {
     use super::*;
     use crate::event::Event;
@@ -3207,300 +2149,6 @@ mod investigation_phase_tests {
     }
 }
 
-/// Handler for [`PlayerAction::DrawEncounterCard`]. Validates phase
-/// + cursor; delegates to [`mythos_draw_for`] on success.
-pub(super) fn draw_encounter_card(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-) -> EngineOutcome {
-    if state.phase != Phase::Mythos {
-        return EngineOutcome::Rejected {
-            reason: format!(
-                "DrawEncounterCard: only valid during Mythos phase, got {:?}",
-                state.phase,
-            )
-            .into(),
-        };
-    }
-    match state.mythos_draw_pending {
-        None => EngineOutcome::Rejected {
-            reason: "DrawEncounterCard: no draw pending (all investigators have drawn)".into(),
-        },
-        Some(expected) if expected != investigator => EngineOutcome::Rejected {
-            reason: format!(
-                "DrawEncounterCard: out of order; expected {expected:?}, got {investigator:?}",
-            )
-            .into(),
-        },
-        Some(_) => mythos_draw_for(state, events, investigator),
-    }
-}
-
-/// Resolves one investigator's full Mythos encounter draw — the
-/// per-card 5-step sub-sequence from Rules Reference p.24, with
-/// surge re-draws looping until the chain ends.
-///
-/// Called by [`draw_encounter_card`] with the pending-drawer's id.
-/// Returns Done on success (chain completed, `mythos_draw_pending`
-/// advanced).
-///
-/// # Mid-chain rejection caveat
-///
-/// `mythos_draw_for` follows the same pattern as `play_card` (CLAUDE.md
-/// documents it): if [`resolve_encounter_card`] rejects mid-chain — e.g.
-/// [`spawn_enemy`] rejecting because the drawing investigator has no
-/// location — the card has already been drawn from `encounter_deck` by
-/// [`draw_encounter_top`], and the apply loop's `events.clear()` on
-/// `Rejected` wipes the event stream but does **not** roll back the state
-/// mutation. The card is silently lost from the encounter deck. In Phase 4
-/// scope this can't happen because the synthetic fixture ensures every
-/// investigator has a location at scenario start; revisit if a future
-/// scenario lets investigators reach a location-less state during play.
-fn mythos_draw_for(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-) -> EngineOutcome {
-    // Fresh chain: count starts at 0 and the loop draws at least one
-    // card (`draw_more = true`).
-    run_mythos_draw_chain(state, events, investigator, 0, true)
-}
-
-/// The Mythos surge-draw loop, shared by the initial draw
-/// ([`mythos_draw_for`]) and the post-suspend resume
-/// ([`resume_spawn_engage`]).
-///
-/// `chain_count` is the surge position already consumed (0 for a fresh
-/// chain); the loop increments it per drawn card and enforces
-/// [`MAX_SURGE_CHAIN`] exactly as the single-pass version did.
-/// `draw_more` gates the first iteration: `true` for a fresh draw,
-/// or the suspended card's surge bit on resume (a non-surge enemy that
-/// suspended for engagement resumes with `draw_more = false`, drawing no
-/// further card — only the cursor advance runs).
-///
-/// On a mid-chain spawn engagement tie, [`resolve_encounter_card`]
-/// returns [`EngineOutcome::AwaitingInput`]; this loop patches the live
-/// `chain_count` into the freshly-stored
-/// [`SpawnEngagePending`](crate::state::SpawnEngagePending) so the resume
-/// continues with the cap budget intact, then returns the suspension.
-///
-/// # Mid-chain rejection caveat
-///
-/// Same as the single-pass version: if [`resolve_encounter_card`]
-/// rejects mid-chain (e.g. [`spawn_enemy`] when the drawing investigator
-/// has no location), the card has already left `encounter_deck` and the
-/// apply loop's `events.clear()` on `Rejected` does not roll back that
-/// mutation. Out of Phase-4 scope (the synthetic fixture gives every
-/// investigator a location at setup).
-pub(super) fn run_mythos_draw_chain(
-    state: &mut GameState,
-    events: &mut Vec<Event>,
-    investigator: InvestigatorId,
-    mut chain_count: usize,
-    mut draw_more: bool,
-) -> EngineOutcome {
-    let Some(reg) = crate::card_registry::current() else {
-        return EngineOutcome::Rejected {
-            reason: "DrawEncounterCard: no card registry installed".into(),
-        };
-    };
-
-    while draw_more {
-        chain_count += 1;
-        if chain_count > MAX_SURGE_CHAIN {
-            unreachable!(
-                "Mythos draw chain exceeded MAX_SURGE_CHAIN ({}) for \
-                 investigator {:?}. Indicates either an infinite reshuffle \
-                 loop (Rules Reference p.18: treachery discard precedes surge \
-                 re-draw, so a surging treachery in a too-small deck cycles \
-                 via the p.10 reshuffle path) or a malformed scenario encounter \
-                 deck. Real scenarios don't surge >{} cards in one chain.",
-                MAX_SURGE_CHAIN, investigator, MAX_SURGE_CHAIN,
-            );
-        }
-
-        // Step 1: Draw the card from the encounter deck.
-        let Some(code) = draw_encounter_top(state, events) else {
-            if chain_count == 1 {
-                return EngineOutcome::Rejected {
-                    reason: "DrawEncounterCard: encounter deck and discard both empty".into(),
-                };
-            }
-            unreachable!(
-                "Mythos draw chain hit empty encounter deck AND empty discard for \
-                 investigator {:?} at chain position {}. Two independent mechanisms \
-                 can reach this: (a) a small encounter deck of only surging \
-                 treacheries can loop infinitely via the Rules Reference p.18/p.10 \
-                 cycle (treachery discard precedes surge re-draw, so the \
-                 just-discarded card gets reshuffled and re-drawn) — caught earlier \
-                 by MAX_SURGE_CHAIN; (b) a small encounter deck of only surging \
-                 enemies exhausts the encounter universe within one chain (enemies \
-                 spawn to play, not discard, so the p.10 reshuffle has nothing to \
-                 pull). Both are scenario-data malformation, not legitimate play.",
-                investigator, chain_count,
-            );
-        };
-
-        let Some(metadata) = (reg.metadata_for)(&code) else {
-            return EngineOutcome::Rejected {
-                reason: format!("DrawEncounterCard: unknown card code: {code:?}").into(),
-            };
-        };
-
-        // Step 2: Check for the peril keyword on the drawn card.
-        skill_test::peril_check(state, events, &code, investigator, metadata.peril);
-
-        // Step 3 + 4: Resolve revelation, then enemy-spawn if applicable.
-        let outcome = resolve_encounter_card(state, events, investigator, code.clone(), metadata);
-        match outcome {
-            EngineOutcome::Done => {}
-            EngineOutcome::AwaitingInput { .. } => {
-                // A mid-chain spawn engagement tie suspended. Record the
-                // live chain position so the resume keeps counting toward
-                // the cap rather than restarting its budget.
-                if let Some(pending) = state.spawn_engage_pending.as_mut() {
-                    pending.chain_count = chain_count;
-                }
-                return outcome;
-            }
-            EngineOutcome::Rejected { .. } => return outcome,
-        }
-
-        // Step 5: If the drawn card has the surge keyword, loop.
-        draw_more = metadata.surge;
-    }
-
-    // Chain complete — advance the cursor.
-    advance_mythos_draw_pending(state, events);
-    EngineOutcome::Done
-}
-
-/// Advance `state.mythos_draw_pending` after a completed chain. If
-/// a next investigator exists in turn order, set to that id.
-/// Otherwise set to None and open the post-1.4 window.
-fn advance_mythos_draw_pending(state: &mut GameState, events: &mut Vec<Event>) {
-    let current = state
-        .mythos_draw_pending
-        .expect("advance_mythos_draw_pending called only after a successful chain");
-    // Eliminated-skip semantics live in `next_active_investigator_after`.
-    let next = cursor::next_active_investigator_after(state, current);
-    state.mythos_draw_pending = next;
-    if next.is_none() {
-        reaction_windows::open_fast_window(state, events, WindowKind::MythosAfterDraws);
-    }
-}
-
-#[cfg(test)]
-mod mythos_draw_for_tests {
-    use super::*;
-    use crate::state::CardCode;
-    use crate::test_support::{test_investigator, TestGame};
-
-    /// Exercises the early-reject guard for the registry / unknown-card
-    /// checks. Depending on which tests have run in this process:
-    ///
-    /// - `"no card registry installed"` — if no registry has been
-    ///   installed yet in this process.
-    /// - `"unknown card code: ..."` — if another test has installed a
-    ///   registry that doesn't know the synthetic code `"__no_such_card"`.
-    ///
-    /// In both cases the invariant is identical: state is not further
-    /// mutated, the card remains in the encounter deck (the draw was
-    /// either blocked before or after the draw). The deck-length
-    /// assertion allows for the draw-then-reject case (deck shrinks by
-    /// at most one) matching the `encounter_card_revealed_tests` pattern.
-    #[test]
-    fn rejects_when_registry_not_installed_or_unknown_code() {
-        let mut state = TestGame::default()
-            .with_investigator(test_investigator(1))
-            .with_phase(Phase::Mythos)
-            .build();
-        state.turn_order = vec![InvestigatorId(1)];
-        state.mythos_draw_pending = Some(InvestigatorId(1));
-        // Seed the encounter deck with an unknown code so we prove the
-        // reject fires at the registry or unknown-code check, not at the
-        // empty-deck check.
-        state
-            .encounter_deck
-            .push_back(CardCode("__no_such_card".into()));
-        let pre_deck_len = state.encounter_deck.len();
-        let mut events = Vec::new();
-        let outcome = mythos_draw_for(&mut state, &mut events, InvestigatorId(1));
-        match outcome {
-            EngineOutcome::Rejected { reason } => {
-                assert!(
-                    reason.contains("no card registry installed")
-                        || reason.contains("unknown card code"),
-                    "unexpected reject reason: {reason:?}",
-                );
-            }
-            other => panic!("expected Rejected, got {other:?}"),
-        }
-        // Deck must not grow; may shrink by 1 if draw happened before
-        // the unknown-code reject (documented exception matching the
-        // encounter_card_revealed validate-first caveat).
-        assert!(
-            state.encounter_deck.len() <= pre_deck_len,
-            "deck should not grow; expected <= {pre_deck_len}, got {}",
-            state.encounter_deck.len(),
-        );
-    }
-}
-
-#[cfg(test)]
-mod draw_encounter_card_tests {
-    use super::*;
-    use crate::test_support::{test_investigator, TestGame};
-
-    #[test]
-    fn rejects_outside_mythos_phase() {
-        let mut state = TestGame::default()
-            .with_investigator(test_investigator(1))
-            .with_phase(Phase::Investigation)
-            .build();
-        state.mythos_draw_pending = Some(InvestigatorId(1));
-        let mut events = Vec::new();
-        let outcome = draw_encounter_card(&mut state, &mut events, InvestigatorId(1));
-        assert!(matches!(
-            outcome,
-            EngineOutcome::Rejected { reason } if reason.contains("only valid during Mythos")
-        ));
-    }
-
-    #[test]
-    fn rejects_when_no_draw_pending() {
-        let mut state = TestGame::default()
-            .with_investigator(test_investigator(1))
-            .with_phase(Phase::Mythos)
-            .build();
-        state.mythos_draw_pending = None;
-        let mut events = Vec::new();
-        let outcome = draw_encounter_card(&mut state, &mut events, InvestigatorId(1));
-        assert!(matches!(
-            outcome,
-            EngineOutcome::Rejected { reason } if reason.contains("no draw pending")
-        ));
-    }
-
-    #[test]
-    fn rejects_when_out_of_order() {
-        let mut state = TestGame::default()
-            .with_investigator(test_investigator(1))
-            .with_investigator(test_investigator(2))
-            .with_phase(Phase::Mythos)
-            .build();
-        state.mythos_draw_pending = Some(InvestigatorId(1));
-        let mut events = Vec::new();
-        // Inv2 attempts to draw when inv1 is expected.
-        let outcome = draw_encounter_card(&mut state, &mut events, InvestigatorId(2));
-        assert!(matches!(
-            outcome,
-            EngineOutcome::Rejected { reason } if reason.contains("out of order")
-        ));
-    }
-}
-
 #[cfg(test)]
 mod mythos_phase_tests {
     use super::*;
@@ -3698,7 +2346,7 @@ mod mythos_phase_tests {
         state.mythos_draw_pending = Some(InvestigatorId(1));
         let mut events = Vec::new();
 
-        advance_mythos_draw_pending(&mut state, &mut events);
+        encounter::advance_mythos_draw_pending(&mut state, &mut events);
 
         assert_eq!(
             state.mythos_draw_pending,
